q1_article_categorizer/requirements.txt
streamlit>=1.32.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
transformers>=4.37.0
torch>=2.2.0
sentence-transformers>=2.5.0
openai>=1.12.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
umap-learn>=0.5.5
nltk>=3.8.1
requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.66.0
numba>=0.58.0
tbb>=2021.0.0 